+++ src/test/java/repast/model/agrosuccess/AgroSuccessEnvrIntegrationTest.java
@@ -1,6 +1,11 @@
 package repast.model.agrosuccess;
 
-import static org.junit.Assert.*;
+import static me.ajlane.geo.repast.RepastGridUtils.gridValueLayerToArray;
+import static me.ajlane.geo.repast.RepastGridUtils.hashGridValueLayerValues;
+import static me.ajlane.geo.repast.RepastGridUtils.totalGridValueLayerValues;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 import java.io.File;
 import org.apache.log4j.BasicConfigurator;
 import org.apache.log4j.Level;
@@ -13,6 +18,7 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.neo4j.graphdb.GraphDatabaseService;
+import me.ajlane.neo4j.EmbeddedGraphInstance;
 import repast.simphony.context.Context;
 import repast.simphony.context.DefaultContext;
 import repast.simphony.dataLoader.ContextBuilder;
@@ -23,11 +29,13 @@ import repast.simphony.parameter.Parameters;
 import repast.simphony.parameter.ParametersParser;
 import repast.simphony.scenario.ScenarioUtils;
 import repast.simphony.valueLayer.GridValueLayer;
-import static me.ajlane.geo.repast.RepastGridUtils.hashGridValueLayerValues;
-import static me.ajlane.geo.repast.RepastGridUtils.totalGridValueLayerValues;
-import static me.ajlane.geo.repast.RepastGridUtils.gridValueLayerToArray;
-import me.ajlane.neo4j.EmbeddedGraphInstance;
 
+/**
+ * TODO: Add test to check {@link LscapeLayer#FireCount} is incremented
+ *
+ * @author Andrew Lane
+ *
+ */
 public class AgroSuccessEnvrIntegrationTest {
 
   final static Logger logger = Logger.getLogger(AgroSuccessEnvrIntegrationTest.class);
@@ -59,7 +67,7 @@ public class AgroSuccessEnvrIntegrationTest {
     ContextBuilder<Object> builder = new AgroSuccessContextBuilder();
     context = builder.build(context);
 
-    // trigger the Heatbug's @ScheduledMethods to be added to the scheduler
+    // trigger the AgroSuccessContextBuilder's @ScheduledMethods to be added to the scheduler
     logger.debug("Scheduling methods...");
     for (Object agent: context.getObjects(Object.class)) {
       logger.debug(schedule.schedule(agent));
@@ -144,6 +152,20 @@ public class AgroSuccessEnvrIntegrationTest {
 
     assertThat("Lct grid should evolve over time, but was unchanged after 5 time steps",
         initialValues, IsNot.not(IsEqual.equalTo(gridValueLayerToArray(lct))));
+  }
+
+  @Test
+  public void oakAgeShouldEvolveOverTime() {
+    GridValueLayer oakAge = (GridValueLayer) context.getValueLayer(LscapeLayer.OakAge.name());
+    int[][] initialValues = gridValueLayerToArray(oakAge);
+
+    for (int i=0; i<5; i++) {
+      // run 5 timesteps to allow time for some spatial variation to emerge
+      schedule.execute();
+    }
+
+    assertThat("OakAge grid should evolve over time, but was unchanged after 5 time steps",
+        initialValues, IsNot.not(IsEqual.equalTo(gridValueLayerToArray(oakAge))));
   }
 
   @Test